Refactor: share the render helpers across the DOM settings tests
Both settings suites had to mount `DomainSettingsProvider` with a context
probe, and both needed the sub-tabs for the region that owns the capped
facility. Two more suites are about to need the same, so lift the harness
into one module before it is copied a third and fourth time.

`seedSettings` also names the "shared vs own" distinction that was
previously an inline boolean: the same snapshot is either put in a link
(read-only shared-view) or in the viewer's own storage (editable), so the
two modes stay comparable in everything but `isSharedView`.

No behaviour change; the suite is unchanged at 10 passing.
